requirements.txt
alembic==1.13.1
amqp==5.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.6.3
boto3==1.35.26
botocore==1.35.26
cachetools==5.4.0
celery==5.4.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
et-xmlfile==1.1.0
fasteners==0.14.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
future==0.18.2
fuzzywuzzy==0.18.0
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
gunicorn==23.0.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
kombu==5.3.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.25.1
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
packaging==24.1
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydicom==3.0.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2022.1
pyxdg==0.27
rapidfuzz==3.9.0
redis==5.0.8
reportlab==4.2.2
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.1
setuptools==74.0.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.0
wheel==0.44.0
xdg==5.0.0
zipp==1.0.0